# backend/internal/handlers/api_ai.go
/*
 * Modular JWT Authenticated API
 *
 * An example API using JWT imported from a separate file.
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package handlers

import (
	"net/http"

	"github.com/duckviet/gin-collaborative-editor/backend/internal/handlers/interfaces"
	"github.com/gin-gonic/gin"
)

type AIAPI struct {
	aiRun interfaces.AIRunAPIHandler
}

var _ interfaces.AIAPIHandler = (*AIAPI)(nil)

func NewAIAPI(aiRun interfaces.AIRunAPIHandler) *AIAPI {
	return &AIAPI{aiRun: aiRun}
}

// Post /api/v1/ai/runs
// Create AI run and stream assistant output
func (api *AIAPI) CreateAiRun(c *gin.Context) {
	if api.aiRun == nil {
		c.JSON(http.StatusNotImplemented, gin.H{"error": "ai run service unavailable"})
		return
	}

	api.aiRun.CreateRun(c)
}

// Post /api/v1/ai/inline-edit
// Inline edit selected text with AI
func (api *AIAPI) InlineEditAi(c *gin.Context) {
	if api.aiRun == nil {
		c.JSON(http.StatusNotImplemented, gin.H{"error": "ai run service unavailable"})
		return
	}

	api.aiRun.InlineEdit(c)
}

// Patch /api/v1/ai/runs/:run_id/consent
// Provide consent for pending AI tool call
func (api *AIAPI) ProvideAiRunConsent(c *gin.Context) {
	if api.aiRun == nil {
		c.JSON(http.StatusNotImplemented, gin.H{"error": "ai run service unavailable"})
		return
	}

	api.aiRun.ProvideConsent(c)
}