# backend/internal/handlers/api_comment.go
/*
 * Modular JWT Authenticated API
 *
 * An example API using JWT imported from a separate file.
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package handlers

import (
	"net/http"

	dbmodels "github.com/duckviet/gin-collaborative-editor/backend/internal/database/models"
	"github.com/duckviet/gin-collaborative-editor/backend/internal/dto"
	"github.com/duckviet/gin-collaborative-editor/backend/internal/handlers/interfaces"
	"github.com/duckviet/gin-collaborative-editor/backend/internal/service"
	"github.com/gin-gonic/gin"
)

type CommentAPI struct {
	commentService service.CommentService
}

var _ interfaces.CommentAPIHandler = (*CommentAPI)(nil)

// NewCommentAPI creates a new comment API handler
func NewCommentAPI(commentService service.CommentService) *CommentAPI {
	return &CommentAPI{
		commentService: commentService,
	}
}

// Post /api/v1/notes/:note_id/comments
// Create a comment on a note
func (api *CommentAPI) CreateComment(c *gin.Context) {
	noteID := c.Param("note_id")
	
	userVal, ok := c.Get("user")
	if !ok {
		c.JSON(http.StatusUnauthorized, gin.H{"error": "unauthorized"})
		return
	}
	u := userVal.(*dbmodels.User)

	var req dto.ReqCreateComment
	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": "Invalid request body"})
		return
	}

	// Create comment
	comment, err := api.commentService.CreateComment(c.Request.Context(), service.CreateCommentRequest{
		NoteID:   noteID,
		UserID:   u.ID,
		Content:  req.Content,
		ParentID: req.ParentId,
	})
	if err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
		return
	}

	c.JSON(http.StatusCreated, comment)
}

// Delete /api/v1/notes/:note_id/comments/:comment_id/delete
// Delete a comment
func (api *CommentAPI) DeleteComment(c *gin.Context) {
	commentID := c.Param("comment_id")
	
	userVal, ok := c.Get("user")
	if !ok {
		c.JSON(http.StatusUnauthorized, gin.H{"error": "unauthorized"})
		return
	}
	u := userVal.(*dbmodels.User)

	if err := api.commentService.DeleteComment(c.Request.Context(), commentID, u.ID); err != nil {
		if err.Error() == "comment not found" {
			c.JSON(http.StatusNotFound, gin.H{"error": err.Error()})
			return
		}
		if err.Error() == "you don't have permission to delete this comment" {
			c.JSON(http.StatusForbidden, gin.H{"error": err.Error()})
			return
		}
		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
		return
	}

	c.Status(http.StatusNoContent)
}

// Get /api/v1/notes/:note_id/comments/list
// List comments for a note
func (api *CommentAPI) ListComments(c *gin.Context) {
	noteID := c.Param("note_id")

	comments, err := api.commentService.ListCommentsByNoteID(c.Request.Context(), noteID)
	if err != nil {
		if err.Error() == "note not found" {
			c.JSON(http.StatusNotFound, gin.H{"error": err.Error()})
			return
		}
		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
		return
	}

	c.JSON(http.StatusOK, gin.H{
		"comments": comments,
		"total":    len(comments),
	})
}

// Put /api/v1/notes/:note_id/comments/:comment_id
// Update a comment
func (api *CommentAPI) UpdateComment(c *gin.Context) {
	commentID := c.Param("comment_id")
	
	userVal, ok := c.Get("user")
	if !ok {
		c.JSON(http.StatusUnauthorized, gin.H{"error": "unauthorized"})
		return
	}
	u := userVal.(*dbmodels.User)

	var req dto.ReqUpdateComment
	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": "Invalid request body"})
		return
	}

	comment, err := api.commentService.UpdateComment(c.Request.Context(), commentID, u.ID, service.UpdateCommentRequest{
		Content: req.Content,
	})
	if err != nil {
		if err.Error() == "comment not found" {
			c.JSON(http.StatusNotFound, gin.H{"error": err.Error()})
			return
		}
		if err.Error() == "you don't have permission to update this comment" {
			c.JSON(http.StatusForbidden, gin.H{"error": err.Error()})
			return
		}
		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
		return
	}

	c.JSON(http.StatusOK, comment)
}

// Get /api/v1/comment/:comment_id
// Get comment detail by comment ID
func (api *CommentAPI) CommentDetail(c *gin.Context) {
	commentID := c.Param("comment_id")

	comment, err := api.commentService.GetCommentByID(c.Request.Context(), commentID)
	if err != nil {
		if err.Error() == "comment not found" {
			c.JSON(http.StatusNotFound, gin.H{"error": err.Error()})
			return
		}
		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
		return
	}

	c.JSON(http.StatusOK, comment)
}