# backend/openapi/dist.yaml
openapi: 3.0.3
info:
  title: Modular JWT Authenticated API
  version: 1.0.0
  description: An example API using JWT imported from a separate file.
servers:
  - url: 'http://localhost:8080/api/v1'
security:
  - BearerAuth: []
paths:
...
  /ai/runs:
    post:
      summary: Create AI run and stream assistant output
      operationId: createAiRun
      tags:
        - AI
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Req_CreateAIRun'
      responses:
        '200':
          description: Streaming SSE response from AI run
          content:
            text/event-stream:
              schema:
                type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  '/ai/runs/{run_id}/consent':
    patch:
      summary: Provide consent for pending AI tool call
      operationId: provideAiRunConsent
      tags:
        - AI
      security:
        - BearerAuth: []
      parameters:
        - name: run_id
          in: path
          required: true
          schema:
            type: string
          description: AI run ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Req_ProvideAIRunConsent'
      responses:
        '200':
          description: Consent applied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Res_ProvideAIRunConsent'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /ai/inline-edit:
    post:
      summary: Inline edit selected text with AI
      operationId: inlineEditAi
      tags:
        - AI
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Req_InlineEdit'
      responses:
        '200':
          description: Inline edit generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Res_InlineEdit'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  '/notes/{note_id}/comments':
    post:
      summary: Create a comment on a note
      operationId: createComment
      tags:
        - Comment
      security:
        - BearerAuth: []
      parameters:
        - name: note_id
          in: path
          required: true
          schema:
            type: string
          description: Note ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Req_CreateComment'
      responses:
        '201':
          description: Comment created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  '/notes/{note_id}/comments/list':
    get:
      summary: List comments for a note
      operationId: listComments
      tags:
        - Comment
      security:
        - BearerAuth: []
      parameters:
        - name: note_id
          in: path
          required: true
          schema:
            type: string
          description: Note ID
      responses:
        '200':
          description: Comments retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  comments:
                    type: array
                    items:
                      $ref: '#/components/schemas/Comment'
                  total:
                    type: integer
                    description: Total number of comments
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  '/notes/{note_id}/comments/{comment_id}':
    put:
      summary: Update a comment
      operationId: updateComment
      tags:
        - Comment
      security:
        - BearerAuth: []
      parameters:
        - name: note_id
          in: path
          required: true
          schema:
            type: string
          description: Note ID
        - name: comment_id
          in: path
          required: true
          schema:
            type: string
          description: Comment ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Req_UpdateComment'
      responses:
        '200':
          description: Comment updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  '/notes/{note_id}/comments/{comment_id}/delete':
    delete:
      summary: Delete a comment
      operationId: deleteComment
      tags:
        - Comment
      security:
        - BearerAuth: []
      parameters:
        - name: note_id
          in: path
          required: true
          schema:
            type: string
          description: Note ID
        - name: comment_id
          in: path
          required: true
          schema:
            type: string
          description: Comment ID
      responses:
        '204':
          description: Comment deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  '/comment/{comment_id}':
    get:
      summary: Get comment by comment Id
      operationId: commentDetail
      tags:
        - Comment
      security:
        - BearerAuth: []
      parameters:
        - name: comment_id
          in: path
          required: true
          schema:
            type: string
          description: Note ID
      responses:
        '200':
          description: Comment retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Enter your JWT token in the format: Bearer <token>.
        The token can be obtained from the `/auth/login` endpoint.
    EditToken:
      type: apiKey
      in: header
      name: X-Edit-Token
      description: Public edit token for note collaboration
  schemas:
...
    Req_CreateAIRun:
      type: object
      properties:
        workspace_id:
          type: string
          example: workspace-123
        session_id:
          type: string
          example: session-123
        note_id:
          type: string
          example: note-123
        message:
          type: object
          properties:
            role:
              type: string
              enum:
                - user
              example: user
            content:
              type: string
              example: Hãy tóm tắt ghi chú này
          required:
            - content
      required:
        - workspace_id
        - session_id
        - note_id
        - message
    Req_ProvideAIRunConsent:
      type: object
      properties:
        tool_call_id:
          type: string
          example: tool-call-123
        approved:
          type: boolean
          example: true
      required:
        - tool_call_id
        - approved
    Res_ProvideAIRunConsent:
      type: object
      properties:
        ok:
          type: boolean
          example: true
        run_id:
          type: string
          example: run-123
        tool_call_id:
          type: string
          example: tool-call-123
        approved:
          type: boolean
          example: true
      required:
        - ok
        - run_id
        - tool_call_id
        - approved
    Req_InlineEdit:
      type: object
      properties:
        workspace_id:
          type: string
          example: workspace-123
        note_id:
          type: string
          example: note-123
        action:
          type: string
          example: improve
        selected_text:
          type: string
          example: this is a draft sentence
        custom_prompt:
          type: string
          example: make this more concise
        context_blocks:
          type: array
          items:
            type: object
            additionalProperties: true
      required:
        - workspace_id
        - action
        - selected_text
    Res_InlineEdit:
      type: object
      properties:
        text:
          type: string
          example: This is a polished sentence.
      required:
        - text
    Comment:
      type: object
      properties:
        id:
          type: string
          example: comment-123
        note_id:
          type: string
          example: note-456
        user_id:
          type: string
          example: user-789
        user_name:
          type: string
          example: John Doe
        user_avatar:
          type: string
          example: 'https://example.com/avatar.png'
        parent_id:
          type: string
          nullable: true
          example: comment-111
        content:
          type: string
          example: This is a comment
        created_at:
          type: string
          format: date-time
          example: '2021-01-01T00:00:00Z'
        updated_at:
          type: string
          format: date-time
          example: '2021-01-01T00:00:00Z'
        replies:
          type: array
          items:
            $ref: '#/components/schemas/Comment'
          example: []
      required:
        - id
        - note_id
        - user_id
        - user_name
        - content
        - created_at
        - updated_at
    Req_CreateComment:
      type: object
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 1000
          example: This is my comment
        parent_id:
          type: string
          nullable: true
          example: comment-111
      required:
        - content
    Req_UpdateComment:
      type: object
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 1000
          example: Updated comment content
      required:
        - content
...
  responses:
    BadRequest:
      description: Bad Request. The request is invalid.
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Invalid request.
    Unauthorized:
      description: 'Unauthorized. The provided JWT token is missing, invalid, or expired.'
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Authentication token is required.
    NotFound:
      description: Not Found. The requested resource was not found.
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Resource not found.
    InternalServerError:
      description: Internal Server Error. The server encountered an unexpected condition that prevented it from fulfilling the request.
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Internal server error.
    Forbidden:
      description: Forbidden - You don't have permission to access this resource
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Forbidden
...